# Optimized Decoding For runtime