Hi! I took the liberty of improving a bit on the runtime, although I must admit there was hardly anything I could improve..
The benchmark I checked in went from: 
```
1 loops: 411.979349 ms
1 loops: 311.466364 ms
```
to 
```
1 loops: 402.03062 ms
1 loops: 236.654808 ms
```

Hope you like it, and if you have any requests, or suggestions, don't hesitate to contact!